refactor: #455 move Copier to utils directory
We do this to mimic the same structure we have in the JS Toolkit generator
because this code has been copied from there and we may want to coalesce it in
the future, when we find a proper place to put it and share it from there.
