+++ packages/generator-liferay-theme/generators/themelet/index.js
@@ -6,8 +6,8 @@
 const path = require('path');
 const Generator = require('yeoman-generator');
 
-const Copier = require('../../lib/Copier');
 const {normalizeName, promptWithQA, sayHello} = require('../../lib/util');
+const Copier = require('../../lib/utils/Copier');
 const versions = require('../../lib/versions');
 
 /**
