+++ packages/generator-liferay-theme/lib/utils/Copier.js
